\name{rgutenberg-package}
\alias{rgutenberg-package}
\docType{package}
\title{A Package for Scraping a Random Project Gutenberg Book}
\description{
  .
}
\details{
\tabular{ll}{
  Package: \tab rgutenberg\cr
  Type: \tab Package\cr
  License: \tab BSD 2-Clause\cr
}
}
\references{
  Project Gutenberg: \url{http://gutenberg.org/}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
