\name{mariner}
\alias{mariner}

\docType{data}

\title{The Rime of the Ancient Mariner}

\description{
  The complete text of Samuel Taylor Coleridge's "Rime of the 
  Ancient Mariner", my favorite poem.
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab The Rime of the Ancient Mariner \cr
  Author:   \tab Samuel Taylor Coleridge \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/151.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(mariner)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(mariner, with.license=FALSE)}.
}

