\name{mississippi}
\alias{mississippi}

\docType{data}

\title{Paradise Lost}

\description{
  The complete text of Samuel Clemens' (Mark Twain's) "Life on the
  Mississippi".
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab Life On The Mississippi, Complete \cr
  Author:   \tab Mark Twain (Samuel Clemens) \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/245.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(mississippi)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(mississippi, with.license=FALSE)}.
}

