\name{shakespeare}
\alias{shakespeare}

\docType{data}

\title{The Complete Works of William Shakespeare}

\description{
  The complete works of William Shakespeare.
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab The Complete Works of William Shakespeare \cr
  Author:   \tab William Shakespeare \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/cache/epub/100/pg100.txt}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(shakespeare)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(shakespeare, with.license=FALSE)}.
}

