\name{bartleby}
\alias{bartleby}

\docType{data}

\title{Bartleby the Scrivener}

\description{
  The complete text of Bartleby the Scrivener.
  
  "I would perefer not to."

}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab Bartleby the Scrivener \cr
  Author:   \tab Herman Melville \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/11231.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(bible)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(bible, with.license=FALSE)}.
}

