\name{beowulf}
\alias{beowulf}

\docType{data}

\title{Beowulf}

\description{
  The complete text of Beowulf.
  
  "Hw\U00E6t!"

}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab Beowulf \cr
  Author:   \tab Unknown \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/16328.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(bible)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(bible, with.license=FALSE)}.
}

