\name{frankenstein}
\alias{frankenstein}

\docType{data}

\title{Frankenstein}

\description{
  The complete text of "Frankenstein".
  
  "Satan has his companions, fellow-devils, to admire and encourage him; but 
  I am solitary and detested."
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab Frankenstein \cr
  Author:   \tab Mary Wollstonecraft (Godwin) Shelley \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/84.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(bible)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(bible, with.license=FALSE)}.
}

