\name{paradise}
\alias{paradise}

\docType{data}

\title{Paradise Lost}

\description{
  The complete text of John Milton's "Paradise Lost".
  
  "Better to reign in Hell, then serve in Heav'n."
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab Paradise Lost \cr
  Author:   \tab John Milton \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/20.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(paradise)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(paradise, with.license=FALSE)}.
}

