\name{inferno}
\alias{inferno}

\docType{data}

\title{Inferno}

\description{
  The complete text of Dante's "Inferno".
  
  "All hope abandon ye who enter here."
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab Inferno \cr
  Author:   \tab Dante Alighieri \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/8789.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(bible)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(bible, with.license=FALSE)}.
}

