\name{occurrence}
\alias{occurrence}

\docType{data}

\title{An Occurrence at Owl Creek Bridge}

\description{
  The complete text of "An Occurrence at Owl Creek Bridge".
  
  "Death is a dignitary who when he comes announced is to be received 
  with formal manifestations of respect, even by those most familiar 
  with him."
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab An Occurrence at Owl Creek Bridge \cr
  Author:   \tab Ambrose Bierce \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/375.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(mississippi)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(mississippi, with.license=FALSE)}.
}

