\name{odyssey}
\alias{odyssey}

\docType{data}

\title{The Odyssey}

\description{
  The complete text of Homer's "The Odyssey".
  
  "Sons are seldom as good men as their fathers; they are generally 
  worse, not better."
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab The Odyssey \cr
  Author:   \tab Homer \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/1727.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(mississippi)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(mississippi, with.license=FALSE)}.
}

