\name{republic}
\alias{republic}

\docType{data}

\title{The Republic}

\description{
  The complete text of Plato's "The Republic".
  
  "I am the wisest man alive, for I know one thing, and that is that I 
  know nothing."
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab The Republic \cr
  Author:   \tab Plato \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/1497.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(paradise)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(paradise, with.license=FALSE)}.
}

