\name{timemachine}
\alias{timemachine}

\docType{data}

\title{The Time Machine}

\description{
  The complete text of H. G. Wells' "Time Machine".

  "Looking at these stars suddenly dwarfed my own troubles and all the 
  gravities of terrestrial life." 
}
\format{
  A \code{pg.book} object.
}
\details{
\tabular{ll}{
  Title:    \tab The Time Machine \cr
  Author:   \tab H. G. (Herbert George) Wells \cr
  Language: \tab English \cr
  URL:      \tab \url{http://www.gutenberg.org/ebooks/35.txt.utf-8}
}
  
  To access the full text unmodified from its Project Gutenberg
  source, call \code{text(shakespeare)}.  To remove Project Gutenberg
  insertions (header information, the license, etc.), call
  \code{text(shakespeare, with.license=FALSE)}.
}

