\name{print}
\docType{methods}

\alias{print-method}
\alias{print,book-method}
\alias{print}

\title{ print }
\description{
  Print method for a the book class.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "book")}}{}
}
}
\usage{
\S4method{print}{book}(x)
}
\arguments{
  \item{x}{book object}
}
\details{
  It prints.
}
\keyword{Methods}
