\name{ rgutenberg }

\alias{rgutenberg}

\title{ Random Book from Project Gutenberg }

\description{
  This function gets a random book from Project Gutenberg and returns
  and object of class 'book'.
}
\usage{
  rgutenberg()
}
\arguments{
}
\details{
  .
}
\value{
  Returns a 'book' class object.
}
\seealso{
  \code{\link{book-class}}
}
\examples{
\dontrun{
library(rgutenberg, quietly=TRUE)
rgutenberg()
}
}
\keyword{Data Generation}


